09_LLMs/MY_RESUME_AI_AGENT/requirements.txt

openai
gradio
pypdf
python-dotenv
requests
